+++ CHANGELOG.md
@@ -8,6 +8,10 @@ Changelog rules:
 -->
 
 ## Unreleased
+- [**BC**] Require php 8.2
+- Update dependencies
+- Add `sprintf` function with a `%A` placeholder for `stringify` function
+- Mark `stringify` constant as deprecated in favor of `sprintf(...)` syntax
 
 ## 6.0.0 - 2023-12-13
 - [**BC**] Require php 8.1